Api index (#10)
* updated seeds file

* WIP request spec for index endpoint

* request spec for index endpoint green

* injectiong the helper path is not possible with js file in the asstes folder
the upside with moving it to the view folder is not worth it. abandoning the idea
cleans up redundant view and controller action. all tests green

* refactors expected resource and adds second expectation
