+++ db/migrate/20160804225744_add_capacity_and_schedule_to_resource.rb
@@ -1,6 +1,6 @@
 class AddCapacityAndScheduleToResource < ActiveRecord::Migration[5.0]
   def change
-    add_column :resources, :capacity, :integer
-    add_column :resources, :schedule, :text
+    add_column :api, :capacity, :integer
+    add_column :api, :schedule, :text
   end
 end
