+++ db/seeds.rb
@@ -11,17 +11,17 @@ description = 'Lorem ipsum dolor sit amet, consectetur adipisicing elit.
 
 Resource.all.each { |r| r.destroy }
 Resource.create(designation: 'Lilla konferanssalen 1', description: description, capacity: 8,
-                schedule: IceCube::Schedule.new(Date.today, duration: 1.day))
+                schedule: IceCube::Schedule.new(Date.today - 1.year, duration: 1.year))
 Resource.create(designation: 'Lilla konferanssalen 2', description: description, capacity: 8,
-                schedule: IceCube::Schedule.new(Date.today, duration: 1.day))
+                schedule: IceCube::Schedule.new(Date.today - 1.year, duration: 1.year))
 Resource.create(designation: 'Stora konferanssalen 1', description: description, capacity: 20,
-                schedule: IceCube::Schedule.new(Date.today, duration: 1.day))
+                schedule: IceCube::Schedule.new(Date.today - 1.year, duration: 1.year))
 Resource.create(designation: 'Stora konferanssalen 2', description: description, capacity: 20,
-                schedule: IceCube::Schedule.new(Date.today, duration: 1.day))
+                schedule: IceCube::Schedule.new(Date.today - 1.year, duration: 1.year))
 Resource.create(designation: 'Stora kongresshallen', description: description, capacity: 200,
-                schedule: IceCube::Schedule.new(Date.today, duration: 1.day))
+                schedule: IceCube::Schedule.new(Date.today - 1.year, duration: 1.year))
 Resource.create(designation: 'Lilla kongresshallen', description: description, capacity: 100,
-                schedule: IceCube::Schedule.new(Date.today, duration: 1.day))
+                schedule: IceCube::Schedule.new(Date.today - 1.year, duration: 1.year))
 
 Resource.all.each do |resource|
   resource.schedule.add_recurrence_rule IceCube::Rule.daily
